commit ada555198cd80127358e38d852840fe349420739
Author: bedalton <Dsbadal@gmail.com>
Date:   Sat Aug 6 10:27:28 2022 -0700

    Add album icon

commit 96f6f5d77e2cad8cb448c46c102232b6ed07c44c
Author: bedalton <Dsbadal@gmail.com>
Date:   Sat Aug 6 10:27:19 2022 -0700

    Update project files

commit 78dec680e4cb65939cb42c255cc3bafb069b531a
Author: bedalton <Dsbadal@gmail.com>
Date:   Sat Aug 6 10:26:46 2022 -0700

    Update Agent-Util to comply with changes in V0.04 package changes

commit 947e39ff79f25dfe1c5de034776bf8228795d49b
Author: bedalton <Dsbadal@gmail.com>
Date:   Sat Aug 6 10:25:02 2022 -0700

    Add additional SPR formats

+++ src/main/java/com/badahori/creatures/plugins/intellij/agenteering/bundles/cobs/decompiler/readC1Cob.kt
@@ -37,7 +37,7 @@ internal fun ByteStreamReader.readC1Cob(fileName:String?): CobBlock.AgentBlock {
     val pictureHeight = int32
     skip(2)
     val image = if (pictureWidth > 0 && pictureHeight > 0)
-        SprSpriteFile.parseFrame(this, position(), pictureWidth, pictureHeight, transparentBlack).flipY().toAwt()
+        SprSpriteFile.parseFrame(this, position().toLong(), pictureWidth, pictureHeight, transparentBlack).flipY().toAwt()
     else
         null
     val agentName = string(uInt8)
